[Docker] Setup basic docker file (#2)
* Move the package.json for the server into the server folder

* Create the docker file for the client

* Create the docker file for the server

* Fix error in server docker file

* Create docker compose for combining client & server docker files

* Create a docker file for building with apache instead of nginx

* Update the server docker container's time zone to be EST

* Create single docker file that bundles the frontend into the backend
The backend now serves the frontend rather than trying to proxy the requests between them. Note: based on the current set up in the server/app.js file, I believe it will throw an error when launching not in docker (like the build folder not existing), but there is also a chance it only happens if trying to call the base route without the build in it. Either way that is something to fix after getting this working (for now it works except the different page routes don't work).

* Fix issue where only main page can be received

* Remove the docker compose files
This had the flaw of the FE not being able to make requests to the BE (with a proxy setup).

* Add the node environment variables that were in the previous docker files

* Delete old apache docker file